moves updates into shortlinkshare component file
