+++ resources/assets/components/utilities/ShortLinkShare/ShortLinkShareContainer.js
@@ -1,7 +1,6 @@
 import { connect } from 'react-redux';
 
-// import ShortLinkShare from './ShortLinkShare';
-import ShortLinkShareUpdate from './ShortLinkShareUpdate';
+import ShortLinkShare from './ShortLinkShare';
 import { getUserId } from '../../../selectors/user';
 
 /**
@@ -15,4 +14,4 @@ const mapStateToProps = state => ({
 });
 
 // Export the container component.
-export default connect(mapStateToProps)(ShortLinkShareUpdate);
+export default connect(mapStateToProps)(ShortLinkShare);
